@@ -10,43 +10,40 @@
   },
   "homepage": "./",
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "1.2.34",
-    "@fortawesome/free-brands-svg-icons": "5.13.0",
-    "@fortawesome/free-solid-svg-icons": "5.15.2",
-    "@fortawesome/react-fontawesome": "0.1.14",
-    "@reduxjs/toolkit": "1.2.5",
-    "@testing-library/jest-dom": "4.2.4",
-    "@testing-library/react": "9.3.2",
-    "@testing-library/user-event": "7.1.2",
-    "@types/jest": "24.0.0",
-    "@types/lodash.debounce": "4.0.6",
+    "@fortawesome/fontawesome-svg-core": "6.2.0",
+    "@fortawesome/free-brands-svg-icons": "6.2.0",
+    "@fortawesome/free-solid-svg-icons": "6.2.0",
+    "@fortawesome/react-fontawesome": "0.2.0",
+    "@reduxjs/toolkit": "^1.8.5",
+    "@types/jest": "29.0.0",
+    "@types/lodash.debounce": "4.0.7",
     "@types/node": "12.0.0",
-    "@types/react": "16.9.0",
-    "@types/react-dom": "16.9.0",
-    "@types/react-redux": "7.1.7",
-    "@types/react-router-dom": "5.1.5",
-    "@types/react-select": "4.0.15",
+    "@types/react": "16.14.31",
+    "@types/react-dom": "18.0.6",
+    "@types/react-redux": "^7.1.24",
+    "@types/react-router-dom": "5.3.3",
+    "@types/react-select": "5.0.1",
     "aspect.js": "0.8.1",
     "cross-env": "7.0.2",
     "electron-is-dev": "1.2.0",
-    "eslint": "6.8.0",
-    "eslint-plugin-jest-dom": "3.6.5",
-    "eslint-plugin-react": "7.19.0",
-    "eslint-plugin-testing-library": "3.10.1",
-    "framer-motion": "1.10.3",
+    "eslint": "8.23.0",
+    "eslint-plugin-jest-dom": "4.0.2",
+    "eslint-plugin-react": "7.31.8",
+    "eslint-plugin-testing-library": "5.6.2",
+    "framer-motion": "7.3.2",
     "fs.extra": "1.3.2",
     "immutable": "4.0.0-rc.12",
     "intl-messageformat-parser": "6.0",
     "lodash.debounce": "4.0.8",
-    "react": "16.13.1",
-    "react-archer": "2.0.2",
-    "react-dom": "16.13.1",
-    "react-intl": "5.7.1",
-    "react-redux": "7.2.0",
-    "react-router-dom": "5.2.0",
-    "react-scripts": "3.4.4",
-    "react-select": "4.3.0",
-    "typescript": "3.8.2"
+    "react": "18.2.0",
+    "react-archer": "4.1.0",
+    "react-dom": "18.2.0",
+    "react-intl": "6.1.1",
+    "react-redux": "^8.0.2",
+    "react-router-dom": "6.3.0",
+    "react-scripts": "5.0.1",
+    "react-select": "5.4.0",
+    "typescript": "4.8.2"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -86,19 +83,9 @@
     ]
   },
   "devDependencies": {
-    "@storybook/addon-actions": "6.2.7",
-    "@storybook/addon-essentials": "6.2.7",
-    "@storybook/addon-links": "6.2.7",
-    "@storybook/addon-viewport": "6.2.7",
-    "@storybook/addons": "6.2.7",
-    "@storybook/node-logger": "6.2.7",
-    "@storybook/preset-create-react-app": "3.1.7",
-    "@storybook/react": "6.2.7",
-    "@storybook/theming": "6.2.7",
-    "concurrently": "5.3.0",
+    "concurrently": "7.4.0",
     "electron": "10.1.0",
     "electron-builder": "22.8.0",
-    "storybook-addon-designs": "5.4.5",
     "wait-on": "5.2.0"
   }
 }
