@@ -1,6 +1,7 @@
 import React, { FunctionComponent } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { globalSelectors, globalSlice } from "../../../slices/GlobalSlice";
+import { useAppDispatch } from "../../../store";
 import { DavitNotification, NotificationLevel } from "./DavitNotification";
 
 export interface ErrorNotificationProps {
@@ -8,7 +9,7 @@ export interface ErrorNotificationProps {
 
 export const ErrorNotification: FunctionComponent<ErrorNotificationProps> = () => {
         const errorMessages: string[] = useSelector(globalSelectors.selectGlobalErrorState);
-        const dispatch = useDispatch();
+        const dispatch = useAppDispatch();
 
         const buildMessage = (message: string, index: number): JSX.Element => {
             return <DavitNotification key={index}
