@@ -2,6 +2,7 @@ import React, { FunctionComponent } from "react";
 import { useDispatch } from "react-redux";
 import { EditActions } from "../../../../../../slices/EditSlice";
 import { GlobalActions } from "../../../../../../slices/GlobalSlice";
+import { useAppDispatch } from "../../../../../../store";
 import { DavitDeleteButton } from "../../../../../atomic/buttons/DavitDeleteButton";
 import { DavitDownloadButton } from "../../../../../atomic/buttons/DavitDownloadButton";
 import { DavitUploadButton } from "../../../../../atomic/buttons/DavitUploadButton";
@@ -13,7 +14,7 @@ export interface ControlPanelFileControllerProps {
 
 export const ControlPanelFileController: FunctionComponent<ControlPanelFileControllerProps> = () => {
 
-    const dispatch = useDispatch();
+    const dispatch = useAppDispatch();
 
     const deleteLocalStorage = () => {
         dispatch(EditActions.setMode.view());
