@@ -35,7 +35,8 @@ export const DataAccess = {
             dataStore.storeFileData(fileData);
             return {...response, code: 200};
         } catch (error) {
-            return {...response, message: error.message};
+            const message = error instanceof Error ? error.message : '';
+            return {...response, message: message};
         }
     },
 
@@ -49,7 +50,8 @@ export const DataAccess = {
             dataStore.createNewProject();
             return {...response, code: 200};
         } catch (error) {
-            return {...response, message: error.message};
+            const message = error instanceof Error ? error.message : '';
+            return {...response, message: message};
         }
     },
 
@@ -64,7 +66,8 @@ export const DataAccess = {
             dataStore.downloadData(projectName);
             return {...response, code: 200};
         } catch (error) {
-            return {...response, message: error.message};
+            const message = error instanceof Error ? error.message : '';
+            return {...response, message: message};
         }
     },
 
@@ -319,7 +322,7 @@ function makeTransactional<T>(callback: () => T): DataAccessResponse<T> {
         dataStore.commitChanges();
     } catch (error) {
         console.warn(error);
-        response.message = error.message;
+        response.message = error instanceof Error ? error.message: "";
         dataStore.roleBack();
     }
     return response;
