@@ -1,12 +1,11 @@
-import { configureStore, getDefaultMiddleware, ThunkAction } from "@reduxjs/toolkit";
+import { Action, configureStore, getDefaultMiddleware, ThunkAction } from "@reduxjs/toolkit";
+import { useDispatch } from "react-redux";
 import { storageMiddleware } from "./middlewares/StateSync";
 import { EditReducer } from "./slices/EditSlice";
 import { globalReducer } from "./slices/GlobalSlice";
 import { MasterDataReducer } from "./slices/MasterDataSlice";
 import { SequenceModelReducer } from "./slices/SequenceModelSlice";
 import { createStorageListener } from "./utils/StorageListener";
-
-const middleware = getDefaultMiddleware().concat(storageMiddleware);
 
 export const store = configureStore({
     reducer: {
@@ -15,10 +14,12 @@ export const store = configureStore({
         edit: EditReducer,
         sequenceModel: SequenceModelReducer,
     },
-    middleware,
+    middleware: (getDefaultMiddleware) => getDefaultMiddleware().concat(storageMiddleware),
 });
 
 window.addEventListener("storage", createStorageListener(store));
 
 export type RootState = ReturnType<typeof store.getState>;
-export type AppThunk = ThunkAction<void, RootState, unknown, any>;
+export type AppThunk = ThunkAction<void, RootState, unknown, Action>;
+export type AppDispatch = typeof store.dispatch;
+export const useAppDispatch = ()=>useDispatch<AppDispatch>();
